% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnp46A_to_wgs84.R
\name{get_vnp46A_extent}
\alias{get_vnp46A_extent}
\title{get_vnp46A_extent}
\usage{
get_vnp46A_extent(path)
}
\arguments{
\item{path}{\code{character} path to hd5 black marble product (VNP46A1 or VNP46A2)}
}
\value{
A \code{numeric} vector containing new extent in correct extent format for `SpatRaster` class ([terra] package)
}
\description{
get_vnp46A_extent
}
