% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic_by_doy.R
\name{mosaic_vnp46a_by_doy}
\alias{mosaic_vnp46a_by_doy}
\title{Mosaic by day of year}
\usage{
mosaic_vnp46a_by_doy(ras_list, output = NULL)
}
\arguments{
\item{ras_list}{named list of `SpatRaster`s. Names must follow NASA VNP46A file naming convention}
}
\value{
nested list where each level contains all images for doy.
}
\description{
Mosaic by day of year
}
\examples{
\dontrun{
 library(terra)

 h5_filenames <- list.files("raster/h5/")
 h5_paths <- list.files("raster/h5/", full.names = T)
 black_marble <- h5_paths |>
    map(~vnp46A_to_wgs84(path =.x)) |>
     set_names(h5_filenames)

 black_marble_by_doy<- mosaic_vnp46a_by_doy(vnp46a1_wgs84)


}
}
