% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnp46A_to_wgs84.R
\name{vnp46A_to_wgs84}
\alias{vnp46A_to_wgs84}
\title{vnp46A_to_wgs84}
\usage{
vnp46A_to_wgs84(path)
}
\arguments{
\item{path}{\code{character} path to hd5 black marble product (VNP46A1 or VNP46A2)}
}
\value{
VNP46A raster/stack in WGS84
}
\description{
project VNP46A black marble products to WGS84. Input can be multi-band. Can use in conjunction with
[purrr::map] to reproject all bands of all rasters contained in folder
}
\examples{
\dontrun{
   library(terra)
   vnp46a1_wgs84 <- vnp46A_to_wgs84(path = "../../raster/VNP46A1.A2022043.h21v04.001.2022044083115.h5")
   # DNB band is 5th band so we can just write that out or we could combine bands however we want
   vnp46a1_dnb <- vnp46a1_wgs84[[5]]
   writeRaster(vnp46a1_dnb,"vnp46a1_dnb.tiff")
}
}
